trading_bot/requirements.txt
ccxt==4.3.25
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
